+++ RollbarAUL/Sources/RollbarAUL/include/RollbarAulLogLevelConverter.h
@@ -24,15 +24,15 @@ NS_ASSUME_NONNULL_BEGIN
 
 //API_AVAILABLE(macos(10.15))
 //API_UNAVAILABLE(ios, tvos, watchos)
-API_AVAILABLE(macos(10.11), ios(9.0), watchos(2.0), tvos(9.0))
+API_AVAILABLE(macos(10.11), ios(9.0), tvos(9.0), watchos(2.0))
 @interface RollbarAulLogLevelConverter : NSObject
 
 + (os_log_type_t) RollbarLevelToAulLevel:(RollbarLevel)value
-API_AVAILABLE(macos(10.11), ios(9.0), watchos(2.0), tvos(9.0))
+API_AVAILABLE(macos(10.11), ios(9.0), tvos(9.0), watchos(2.0))
 ;
 
 + (RollbarLevel) RollbarLevelFromAulLevel:(os_log_type_t)value
-API_AVAILABLE(macos(10.11), ios(9.0), watchos(2.0), tvos(9.0))
+API_AVAILABLE(macos(10.11), ios(9.0), tvos(9.0), watchos(2.0))
 ;
 
 @end